requirements.txt
arrow==1.3.0
birdsong @ git+https://github.com/CorsoSource/birdsong.git@304074c68c480a075ec1f25f58766291db3f0e3b
certifi==2025.4.26
charset-normalizer==3.4.2
ciso8601==2.3.2
idna==3.10
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
keyring==25.6.0
more-itertools==10.7.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
ruff==0.11.11
six==1.17.0
types-python-dateutil==2.9.0.20250516
urllib3==2.4.0
